% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot Location Data on State Map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A unique identifying number for each state (FIPS code). An
error will be generated if a number is assigned to this parameter which
does not exist within the data for the specified year. In the case that no
accident data exists for the specified state and year, the message "no
accidents to plot" will be generated. Values of class integer or numeric
may be safely assigned to this parameter. Values of class character
are also safe if coercible to integer, e.g. "48". If value is not
coercible to integer, invalid results may occur due to introduction of NAs
by coercion or an error may be generated.}

\item{year}{A value indicating the year of the desired data. Values of class
integer or numeric may be safely assigned to this variable. Values of class
character are also safe if coercible to integer, e.g. '2013'. If the
value is not coercible to integer, invalid results may occur due to
introduction of NAs by coercion or an error may be generated.}
}
\value{
Returns NULL.
}
\description{
This function extracts data from a file of accident records for a specified
year and displays a simple map of the desired state with a dot plotted at
the location of every accident recorded in the data for that state during
the specified year.
}
\examples{
\dontrun{
fars_map_state(48, 2013)
fars_map_state(6, "2015")
}

}
