% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read Data from File}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the name of the file to be read.
An error is generated if no file with the specified name exists in the
working directory.}
}
\value{
A data.frame object containing the contents of the specified file.
}
\description{
This function checks for the existence of a file having a specified name.
If the file exists, its contents are read and returned by the function.
}
\note{
To ensure a filename passed to this function is properly constructed
use the function in this package called make_filename.
}
\examples{
\dontrun{fars_read("accident_2013.csv.bz2")
fars_read(make_filename(2013))
}

}
