% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize a Year's Data as Monthly Totals}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A value or vector/list of values indicating the year(s) of the
data desired. Values of class integer or numeric may be passed safely to
this function. Values of class character are also safe if coercible to
integer, e.g. '2013'. Data only exist for the years 2013, 2014, & 2015. An
error will be generated if a year other than one of these is passed to the
function.}
}
\value{
This function returns a data.frame object whose first column is the
sequence of integers from 1 to 12 (enumerating the 12 months of the year).
Each subsequent column represents a summary of the desired years in terms
of monthly totals.
}
\description{
This function extracts data from a file of accident records for specified
years and summarizes the specified year(s) of accident data in a table
of monthly accident totals.
}
\examples{
\dontrun{fars_summarize_years(2015)
fars_summarize_years(c('2013','2014'))
fars_summarize_years(list(2014,2015)) }

}
