% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make Filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A value indicating the year of the desired data. Values of class
integer or numeric may be safely assigned to this variable. Values of class
character are also safe if coercible to integer, e.g. '2013'. If the
value is not coercible to integer, invalid results may occur due to
introduction of NAs by coercion or an error may be generated.}
}
\value{
A character string holding the desired filename. As a side effect,
this function also prints out the character string.
}
\description{
Generates the name of a file containing accident data for a specified year.
}
\note{
Data only exist for the years 2013, 2014, & 2015. However, if another
year is chosen, the filename is still returned with no error generation.
}
\examples{
make_filename(2013)
make_filename("2013")

}
